requirements.txt
fastapi
uvicorn[standard]
python-multipart
redis
python-dotenv